Major Refactor of messaging systems
Previously, the Casium runtime was responsible for maintaining a global
store of state. Even the user has not got Casium DevTools installed,
this state was being maintained, which could be taxing over time for
non-developers using complex applications.

Instead, DevTools should be responsible for maintaining this state. The
majority of Casium's `notify` and `intercept` functions have been
extracted to the `src/common/client.ts` module. When running in a
WebExt context, the compiled version of this client can be executed in
the context of the inspected application using the WebExt's script
injection API.

When running in an Electron context, a compiled version of this client
(with a WebSocket message passing mechanism) can be injected into the
inspected application with a `<script/>` tag.