Remove 'messaging' script and make `panel.html` common
The actual HTML of the Panel should be consistent across
implementations, so it can be moved to the `common` dir. It is assumed
that each implementation will provide its own `panel.js`.

Eventually the hacky `LISTENERS` array on Window will go away, but for
now it can be moved to `web-ext/panel.ts`.