Implement basic Electron main process, renderer and WebSocket server
In order to support the Electron app, three components are required:

1. The `Server` class

This class initializes an HTTP and WebSocket server. It serves
`injected-script` over HTTP, that is intended to be included on the page
to be inspected. This script creates a WebSocket connection to the
Server that invokes and responds to messages dispatched by the core
DevTools Instrumenter.

The Server instance emits events when a WebSocket connection is opened,
closed or a message is received.

2. The Electron Main Process

Initializes the Electron UI frame and creates a Server instance.
Whenever an event is emitted by the Server, it is relayed straight
through to the Renderer using Electron's IPC mechanism.

3. The Electron Renderer

Responsible for the Panel UI itself, creates an `<App/>` component in
the same way as the existing WebExtension. Instead of using the
background page and `window.postMessage` as a messaging channel, it
uses Electron's IPC mechanism.